Adds proof of concept for the jigsaw app
Ugly code and so on, but this seems to verify the approach
of using pillow to generate images that are used as
sprites with pyglet and opengl. Depth testing is enabled
to handle the z-ordering, which requires a modification of
the sprite class. Zooming is handled by modifying the
orthographic projection.

Performant up to at least 10k pieces, although moving multiple
sprites at once is expensive, should be restricted for more than
about 100 pieces. Re-ordering the z-levels is also slightly
expensive and there is a small but noticeable delay when selecting
a piece when there are >10k pieces.

Pillow can cut elliptical jigsaw pieces fast enough, even for very
large images and 10k pieces (startup delay of a few seconds).
Bezier curves will be more expensive, but hopefully not
prohibitively so. Should be possible to make new images when
pieces are connected, but I am yet to test this.