Implements curved pieces
More ugly code. Proof-of-concept for bezier-based pieces. Bottleneck
is sampling the bezier-curves, which should be possible to speede up
by a factor of around 10 by using parallelization and vectorization.
With small number of sample points per edge (40 is default), 10k
pieces can be created in just a few seconds.

Fixed selection delay by using quad-trees for collision detection
(adds dependency to external lib, which is pure python though),
now there is no noticeable delay for selecting pieces with 10k
pieces. With 50k pieces there is a ~0.2s delay due to z-levels.
There is a load-time penalty of a few seconds for the quad-tree
construction, but after that collision detection is incredibly fast.

Pixel-perfect collision detection is gained by looking at the alpha
level of the pixel data of the piece.