Switches from sprite to polygon
Further calculations reveal that pre-processing the pieces
into images is a bad idea, because combining the images
will potentially require prohibitively large amounts of
memory (as much as n^2, where n is the number of pieces).

To remedy this, I switched to just making the cutouts in
OpenGL directly. I found a python implementation of the earcut
algorithm for triangulating non-convex polygons, and it is
reasonably quick, although with worse-than-linear time
complexity in the number of vertices. This becomes prohibitively
expensive around about 2000 vertices per piece, which means
that combining pieces is not an option. Also, moving a
single piece becomes choppy around 3200 vertices.

The solution turned out to be rather easy: just use
glTranslatef in a separate drawing group for held-pieces, then
change the coordinates of the vertices only when the piece is
dropped. This appears to be performant with arbitrary number of
pieces and vertices: tested with 20k pieces sampled at 40 points
per edge, and 1000 pieces sampled at 400 points per edge, and
maintained a cool 1000 fps.

This solution should allow for group-selection of arbitrary number
of pieces without loss of performance, while also looking better
than the pre-rendered sprites. The only real drawback is that
the loading times increased by a factor of about 10, so that 1000
pieces takes about 6 seconds to load. The major bottleneck (about
90%) is due to the earcut algorithm, which gets a lot more expensive
for more vertices. It could probably be sped up about 10x with a
C++ port, and the whole piece generation process should be easy to
parallelize for another 12-fold speedup (on this system).

TODO:
Z-levels, point-in-polygon hit-detection, selection box, cleanup