Work to snap and merge pieces together
Now the pieces are merged, but the hitbox is not updated. Also, the
merging operation moves the vertices from both pieces into a new
vertex_list, which becomes expensive for large pieces. Even just
moving the large pieces is expensive in the final step.

Todo is to create new groups when the pieces are large enough.