Implements MVC architecture and translation groups
Major refactoring. Now the view and model are separated, so that
only the view knows about OpenGL specific things, and the model
takes care of logical things like piece positions, snapping logic
and such. The view also listens to input and will dispatch pyglet
events to the controller, which then either changes or asks the
model. The model in turn may dispatch further events in response,
which are again caught by the controller, which may notify the view.

Moving a piece by updating its vertices is expensive, and changing
the drawing group of a piece is also expensive if there are many
vertices. At the same time, having many drawing groups is also
expensive, so we can't have one for each piece. To solve this, merged
pieces are moved with vertex updates when they are low in number,
but once enough pieces have merged together, they are all moved to
a new TranslationGroup, from which they are never moved. At that
point, all updates to the position of that piece is handled by a
translation matrix, which is fast.

With a threshold of 100 pieces per group, I get at least 300 fps
with 10k pieces, with no noticeable delay on merging.

Oh, there's also a cheat button (c to cheat once, x to cheat 100
times), for testing purposes.