WIP switch to experimental pyglet branch
A lot of small changes in order for shaders and so on to work. Requires
switching from legacy OpenGL, which means I can't use the translation
matrix thing with the groups anymore, which has big impacts on a large
part of the view class. Also I can't use glooey for unknown reasons.