Finishes the transition to modern OpenGL
One solution turned out to be staring me in the face all along, that was
very simple to implement once I knew how. By creating multiple
ShaderProgram objects and associating each one with a Group, I could let
each one write to their own shader uniform. This essentially replaces
the old fixed pipeline translation matrix, so that instead of changing
that matrix, I simply write the proper translation for each group to the
uniform.

Of course this is only easy because I already did the hard parts before.
It would have been a lot easier to have something like a lookup table
uniform for each piece offset, and use a vertex attribute to store the
index in the uniform array. This approach, although simple to implement,
has several drawbacks: First, there's a limit to the size of each
uniform array: 4096 bytes on my system, and there's a total limit for
how much data can be stored in all uniforms combined. This amount is not
enough to store all the offsets for 10k pieces. It's probably possible
to overcome this by using shader storage object buffers (ssob's)
instead, but I didn't try that. The second problem is that writing to
the uniform is not exactly free. I could only move a few hundred pieces
with a smooth framerate with this approach. It might be possible to make
it quicker by writing larger parts of the uniform in one chunk, rather
than having separate opengl calls for each piece, but I haven't tested
that either.

All said and done, it might be feasible to increase performance with
some clever use of SSOBs and other trickeries, but I don't see any
particularly easy way of gaining a large boost. The current approach in
this commit should be approximately as fast as the one that is using the
old opengl.

Of course, if I could create a really good looking pre-rendered image
for each piece, then I should be able to use instanced rendering to
obtain a potentially massive speedup. But I think that is for another
time.