Implements normal mapping for piece shader
Requires pillow to create the normal and height maps, and aggdraw to
draw antialiased lines for creating the height map. Generates the normal
map using code I adapted from pythonstuff.org.

Although it's a big improvement in the visuals, there's still some
things left to do to make it look better. In particular, the normal map
has a thin blue/purple line right between each piece, which allows light
to shine on the very edge of the piece in a way that looks a little bad.
I think if I could smear the height map a little bit, without
necessarily blurring it a lot, or otherwise adapt the normal map filter,
I could obtain better results. It is also possible that parallax mapping
could make the image look more realistic.

Finally, I'm thinking of making a small overlay on the height map with
some paper-like texture, just to get a bit of roughness in there,
mimicking the look of a real jigsaw piece.