Attempts to fix a z-level related bug
Sometimes when connecting small pieces to big pieces, the z-level
of the new piece is using the z-level of the big piece instead of the
small piece, which can lead to a selected pieces being underneath other
pieces, which looks bad. This fixes the bug in some situations, but
there are still problems, especially when using the cheat button.