Adds trays that can be hidden, to put pieces on
There are 10 trays, numbered 0-9, with all pieces being in tray 0 to
begin with. All trays are visible at first, and pressing the keys 0-9
will toggle visibility of the corresponding tray, and all pieces
contained therein. Pieces can be moved to a different tray by selecting
them and pressing Ctrl+# for the correspondingly numbered tray.

Pieces can be snapped together even if they are on different trays -
they will then be moved to the tray of the piece that was held. Pieces
will not snap together with hidden pieces. Hidden pieces can't be moved
or selected. Hiding a tray will drop all held pieces that was in that
tray. Putting a piece in a hidden tray will drop (and hide) that piece,
but there will be no snapping even if the piece was close to a match.

There are still some bugs relating to z-levels that may have changed
slightly as a result of the trays.