Fixes z-level issues
Turns out the MAX_Z_DEPTH must be small enough to distinguish between
individual ints, which puts it roughly around 5000000. This caused one
of the more elusive z-level bugs, in which a big piece on certain
z-levels didn't appear to move to the top when selected.

The other bugs relating to putting pieces together have also hopefully
been solved.