Some bug fixes and other improvements
Clicking a piece and then pressing a number will no longer change the
tray of the piece - instead it is necessary to either select the piece
together with others, or hold it (not releasing the mouse button), or
clicking the piece while holding the number key. This behavior prevents
accidentally changing tray of a piece after simply having moved it out
of the way.