Removes json serialization in favour of pickle
The json serialization turned out to be relatively slow, because I have
to iterate over all the polygons and convert them to json. For this
reason I switched to pickle, which is less secure, but faster and
simpler. A lot of the serialization code can therefore be removed.

This commit implements quicksave (f5) and quickload (f9), and uses the
new package compress_pickle to also compress the saved game state to
save some space. Loading will save all the piece positions, as well as
their trays and which trays are visible, but it does not save the actual
textures used (only the path of the original image), so the normal map
is created again from the polygons.

The save files could be made a lot smaller if I chose to save only the
random seed needed to create the pieces instead of all the points in
the polygons, but the drawback of such a solution would be a longer
loading time, since all the polygons would have to be generated first.
Depending on how much I can optimize that generative process later on,
it might be worth it to make this space optimization.

Regarding the security of using pickles, I doubt it will be an issue
since the program will only load pickles created by that same program.
For added security later on, I could also use cryptography to sign the
files, so that only a program with access to the signature key will be
able to create save files that can be opened by the program. But this
probably only opens up other cans of security issues, most of which are
unlikely to be relevant in any case.