Adds timer and pause key
Timer keeps track of how long you've played a game. It is paused when
the game is paused. Press PERIOD (.) to print the number of seconds the
game has been running. Press PAUSE to pause the game. Press PAUSE again
to continue playing. While paused, all keyboard and mouse inputs are
ignored, except the PAUSE key and CTRL+R for creating a new game. The
puzzle is hidden while the game is paused (screen is black).